% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.critical.T2}
\alias{asymptotic.critical.T2}
\title{asymptotic.critical.T2}
\usage{
asymptotic.critical.T2(m, n, alpha = 0.1)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{alpha}{: significance level. Default value is set equal to 0.1}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_2} test statistic
based on asymptotic normal approximation
\deqn{N\left(\frac{n*(m+n+1)}{2}, \frac{n*m*(n+m+1)}{12}\right).}
}
\description{
It computes the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_2} (Wilcoxon-Mann-Whitney) test statistic
based on asymptotic normal approximation.
}
