% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.critical.T3}
\alias{asymptotic.critical.T3}
\title{asymptotic.critical.T3}
\usage{
asymptotic.critical.T3(m, n, alpha = 0.1)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{alpha}{: significance level. Default value is set equal to 0.1}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_3} test statistic
based on asymptotic normal approximation
\deqn{N \left(\frac{\choose(n,2)*\choose(m,2)*\\theta}{N}+\frac{n*(2*n^2-3*n+1)}{6}+\frac{n*(n-1)}{2},
\frac{\choose(n,2)*\choose(m,2))}{N}*\frac{64}{(45*N*\lambda^3*(1-\lambda)^3)\right)}.}
}
\description{
It computes the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_3} test statistic
based on asymptotic normal approximation.
}
