% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{asymptotic.critical.Tk}
\alias{asymptotic.critical.Tk}
\title{asymptotic.critical.Tk}
\usage{
asymptotic.critical.Tk(m, n, k, alpha = 0.1)
}
\arguments{
\item{m}{calibration sample size.}

\item{n}{test sample size.}

\item{k}{order of the higher order Wilcoxon sum-rank test statistic.}

\item{alpha}{significance level. Default value is set equal to 0.1.}
}
\value{
A number, which is the \eqn{(1-\alpha)}-quantile of higher order Wilcoxon sum-rank test statistic
based on asymptotic normal approximation.
}
\description{
It computes the \eqn{(1-\alpha)}-quantile of the higher order Wilcoxon sum-rank \eqn{T_k} test statistic
based on asymptotic normal approximation.
}
