% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.pvalue.T2}
\alias{asymptotic.pvalue.T2}
\title{asymptotic.pvalue.T2}
\usage{
asymptotic.pvalue.T2(m, n, T.obs)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{T.obs}{: observed value of the test statistic}
}
\value{
It returns the approximated \emph{p}-value of the LMPI \eqn{T_2}
test statistic based on the asymptotic normal approximation
\deqn{N\left(\frac{n*(m+n+1)}{2}, \frac{n*m*(n+m+1)}{12}\right).}
}
\description{
It computes the approximated \emph{p}-value of the LMPI \eqn{T_2} (Wilcoxon-Mann-Whitney)
test statistic based on the asymptotic normal approximation.
}
