% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.pvalue.T3}
\alias{asymptotic.pvalue.T3}
\title{asymptotic.pvalue.T3}
\usage{
asymptotic.pvalue.T3(m, n, T.obs)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{T.obs}{: observed value of the test statistic}
}
\value{
It returns the approximated \emph{p}-value of the LMPI \eqn{T_3}
test statistic based on the asymptotic normal approximation
\deqn{N \bigg(\frac{\choose(n,2)*\choose(m,2)*\theta}{N}+\frac{n*(2*n^2-3*n+1)}{6}+\frac{n*(n-1)}{2},
\frac{\choose(n,2)*\choose(m,2))}{N}*\frac{64}{(45*N*\lambda^3*(1-\lambda)^3)\bigg)}.}
}
\description{
It computes the approximated \emph{p}-value of the LMPI \eqn{T_3}
test statistic based on the asymptotic normal approximation.
}
