% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{asymptotic.pvalue.Tk}
\alias{asymptotic.pvalue.Tk}
\title{asymptotic.pvalue.Tk}
\usage{
asymptotic.pvalue.Tk(m, n, k, T.obs)
}
\arguments{
\item{m}{calibration sample size.}

\item{n}{test sample size.}

\item{k}{order of the higher order Wilcoxon sum-rank test statistic.}

\item{T.obs}{observed value of the test statistic.}
}
\value{
A number, which is the approximated \emph{p}-value of the higher order Wilcoxon sum-rank
test statistic based on the asymptotic normal approximation.
}
\description{
It computes the approximated \emph{p}-value of the LMPI \eqn{T_k}
test statistic based on the asymptotic normal approximation.
}
