% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{compute.critical.values}
\alias{compute.critical.values}
\title{compute.critical.values}
\usage{
compute.critical.values(
  m,
  n,
  local_test = "wmw",
  alpha,
  k = NULL,
  stats_G_vector,
  n_perm = 0,
  B = 10^3,
  critical_values = NULL,
  seed = 123
)
}
\arguments{
\item{m}{calibration size.}

\item{n}{test size.}

\item{local_test}{local test to be used in the closed testing procedure. Default value is Wilcoxon sum-rank test.}

\item{alpha}{significance level.}

\item{k}{order of the higher order Wilcoxon sum-rank test statistic.}

\item{stats_G_vector}{vector of elementary test statistics to perform the test in Shiraishi (1985).}

\item{n_perm}{if \eqn{min(m,n)\leq n_perm} critical values will be computed via permutation. Default value is 0.}

\item{B}{number of permutation to compute critical values. Default value is 10^3.}

\item{critical_values}{if not \code{NULL}, a vector of precomputed critical values obtained using
the permutation distribution of the test statistic.}

\item{seed}{seed to ensure reproducible results.}
}
\value{
A vector of numeric values that are the critical values at significance
level \eqn{\alpha} for a test statistic chosen among higher order Wilcoxon sum-rank,
Fisher or Shiraishi statistics.
}
\description{
It computes the vector of critical values for a chosen test statistic
at significance level \eqn{\alpha}, one for each integer between 1 and the test size.
}
