% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{compute.global.pvalue.shirashi}
\alias{compute.global.pvalue.shirashi}
\title{compute.global.pvalue.shirashi}
\usage{
compute.global.pvalue.shirashi(S_X, S_Y, g, num_mc = 1000)
}
\arguments{
\item{S_X}{calibration score vector.}

\item{S_Y}{test score vector.}

\item{g}{outlier density.}

\item{num_mc}{number of Monte Carlo iterations.}
}
\value{
A number, the global \emph{p}-value for the Shiraishi test.
}
\description{
Given the outlier density function and the calibration and test score vectors,
it computes the global \emph{p}-value for the Shiraishi test.
}
