% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_theta.Tk}
\alias{compute_theta.Tk}
\title{compute_theta.Tk}
\usage{
compute_theta.Tk(m, n, k)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}
}
\value{
It returns the expectation of LMPI \eqn{\left[\frac{T_k} -\sum_{i=1}^n(i-1)^k-1\right]{\binom{m,k-1}\binom{n,k-1}}} in the limit,
computed according to the Central Limit Theorem for \eqn{U}-statistics.
}
\description{
compute_theta.Tk
}
