% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_theta.Tk.tilde}
\alias{compute_theta.Tk.tilde}
\title{compute_theta.Tk.tilde}
\usage{
compute_theta.Tk.tilde(m, n, k, lambda)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}

\item{lambda}{: proportion of calibration observations in the pooled sample of calibration and test set}
}
\value{
It returns the expectation of LMPI \eqn{T_k} in the limit,
computed according to the Central Limit Theorem for \eqn{U}-statistics.
}
\description{
compute_theta.Tk.tilde
}
