% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitneyk3.R
\name{crit.T3}
\alias{crit.T3}
\title{crit.T3}
\usage{
crit.T3(S_X, S_Y, alpha = 0.1, B = 10^3, seed = 123)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}

\item{B}{: number of interation to approximate \eqn{T_3} distribution via permutation}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A R object of class \emph{crit.vals.info}, which is a list consisting
of the following elements: \itemize{
\item m number of test observations
\item n number of calibration observations
\item alpha significance level of the local test
\item crit.val critical value of LMPI \eqn{T_3} statistic.}
}
\description{
Given the number of observations in the calibration set (\eqn{m}) and the number
of observations in the test set (\eqn{n}), it returns the vector of critical
values \eqn{U(m,h)} of the Wilcoxon-Mann-Whitney test statistic letting \eqn{h} vary between
\eqn{1} and \eqn{n}, while the size of the first sample is kept fixed equal to \eqn{m}.
}
\examples{
crits = crit.T3(S_X=runif(10),S_Y=runif(min=0, max=0.5, 5),alpha=0.1)


}
