% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_WMW.R
\name{critWMW}
\alias{critWMW}
\title{critWMW}
\usage{
critWMW(m, n, alpha = 0.1)
}
\arguments{
\item{m}{: number of observations in the test set}

\item{n}{: number of observations in the calibration set}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}
}
\value{
A R object of class \emph{crit.vals.info}, which is a list consisting
of the following elements: \itemize{
\item m: number of test observations
\item n: number of calibration observations
\item alpha: significance level of the local test
\item crit.vals: vector of critical values of the Wilcoxon-Mann-Whitney
statistic \eqn{U(n,k)}, \eqn{k=1,\ldots,m}. Given two samples
\eqn{(X_1,\ldots,X_n)} and \eqn{(Y_1,\ldots,Y_k)}, the Wilcoxon-Mann-Whitney
statistic is computed as \deqn{ U(n,k) = \sum_{i = 1}^{n}\sum_{j = 1}^{k} \mathbb{1}\{X_i > Y_j\}}
and it has mean \eqn{kn/2} and variance \eqn{kn(k+n+1)/12}.

When \eqn{n<1000} and \eqn{m<10} the value of the Wilcoxon-Mann-Whitney statistic
is exactly computed with the R function \emph{\strong{qwilcox}}; otherwise, we use
the normal approximation using continuity correction \deqn{ U(n,k) \sim N(kn/2-1/2, kn(k+n+1)/12)}.
}
}
\description{
Given the number of observations in the calibration set (n) and the number
of observations in the test set (m), it returns the vector of critical
values of the test statistic Wilcoxon-Mann-Whitney letting the size of
the second sample vary between 1 and m, while the size of the first sample
is kept fixed equal to n.
}
\examples{
critWMW(m=100,n=2000)

critWMW(m=100,n=900)

m=10;n=20
crits = critWMW(m=m, n=n)$crit.vals
plot(x=1:m, y=crits, xlab="m-k+1", main = "Critical values of WMW statistic for n=20 and k=1,...,m")



}
