% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.S_t3.R
\name{d.S_t3}
\alias{d.S_t3}
\title{d.S_t3}
\usage{
d.S_t3(S_X, S_Y, S, alpha = 0.1, n.exact = 10, B = 10^3, seed = 123)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{S}{: vector of selected indices in the test set}

\item{alpha}{: significance level. Default level is set equal to 0.1}

\item{n.exact}{: if \eqn{min\{m,n\}\leq n.exact} the critical values of the LMPI \eqn{T_3}
statistic are computed via permutation. Default value is set equal to 10}

\item{B}{: number of iteration to approximate \eqn{T_3} distribution via permutation}

\item{seed}{: seed to ensure reproducible results}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Wilcoxon-Mann-Whitney local test that rejects the global null hypothesis at level \eqn{\alpha}
when the test statistic \deqn{\sum_{i=1}^n\sum_{j=1}^m \mathbb{1}\{X_j<Y_i\}} is greater than
the critical value corresponding to the significance level \eqn{\alpha}.
The selection set can be any arbitrary subset in the index test set
according to which units we want to test.
}
\description{
For any aribitrary selection subset of the index test set, the function returns a confidence
lower bound for the number of true discoveries provided by closed testing procedure using the LMPI \eqn{T_3} local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d.S_t3(S_Y=Sy, S_X=Sx, S=c(1:3, 21), alpha=0.1)



}
