% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.selection_simes.R
\name{d.selection_simes}
\alias{d.selection_simes}
\title{d.selection_simes}
\usage{
d.selection_simes(S_Y, S_X, S, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{S}{: vector of selected indices corresponding to the test observations to test}

\item{alpha}{: significance level. Default values is set equal to 0.1}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Simes local test applied to conformal \emph{p}-values.
The selection set can be any arbitrary subset in the index set of test observations
according to which test units we want to test.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test for any selection in the test index set.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d.selection_simes(S_Y=Sy, S_X=Sx, S = 5:15)

}
