% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.selection_storeysimes.R
\name{d.selection_storeysimes}
\alias{d.selection_storeysimes}
\title{d.selection_storeysimes}
\usage{
d.selection_storeysimes(S_X, S_Y, S, lambda = 0.5, alpha = 0.1)
}
\arguments{
\item{S_X}{vector of calibration scores}

\item{S_Y}{vector of test scores}

\item{S}{set of indices corresponding to the \emph{p}-values to be tested}

\item{lambda}{: parameter involved in the computation of Storey estimator.
Default value is set equal to 0.5}

\item{alpha}{significance level. Default values is set equal to 0.1}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Simes local test with Storey's estimator for the proportion of true null
hypotheses applied to conformal \emph{p}-values.
The selection set can be any arbitrary subset in the index set of test observations
according to which test units we want to test.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test for any selection in the test index set.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d.selection_storeysimes(S_Y=Sy, S_X=Sx, S = 1:29, alpha = 0.6)

}
