% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_BH.R
\name{d_benjhoch}
\alias{d_benjhoch}
\title{d_benjhoch}
\usage{
d_benjhoch(S_Y, S_X, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{alpha}{: significance level. Default level is set equal to 0.1}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in Benjamini-Hochberg procedure applied to.
The selection set, i.e. the set of hypothesis
indices that we are interested in is \eqn{[m]=:\{1,...,m\}} by default.
}
\description{
d_benjhoch
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_benjhoch(S_Y=Sy, S_X=Sx)


}
