% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_choose.method.R
\name{d_choose.method}
\alias{d_choose.method}
\title{d_choose.method}
\usage{
d_choose.method(
  S_X,
  S_Y,
  methods = c("adjustedfisher", "t2", "t3", "simes", "storeysimes"),
  lambda = 0.5,
  n.exact.f = 10,
  n.exact.t2 = 10,
  n.exact.t3 = 10,
  B.f = 10^3,
  B.t3 = 10^3,
  seed = 123,
  alpha = 0.1
)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{methods}{: a character vector indicating one or more of the following local tests to be used in the closed testing procedure:
\itemize{
\item "simes": Simes' test rejects the global null when
\item "storeysimes": Simes' test with Storey estimatore for the proportion of true null hypotheses rejects the global null when ,
\item "adjustedfisher": adjusted Fisher test rejects the global null when ,
\item "t2": LMPI \eqn{T_2} or Wilcoxon-Mann-Whitney test rejects the global null when ,
\item "t3": LMPI \eqn{T_3} test rejects the global null when}
Default value is the vector including all possible local tests.}

\item{lambda}{: parameter involved in the computation of Storey estimator, when method "storeysimes" is used.
Default value is set equal to 0.5}

\item{n.exact.f}{: if \eqn{min\{m,n\}\leq n.exact.f} the critical values of the adjusted Fisher test
statistic are computed via permutation. Default value is set equal to 10}

\item{n.exact.t2}{: if \eqn{min\{m,n\}\leq n.exact.t2} the critical values of the LMPI \eqn{T_2}
statistic are computed via permutation. Default value is set equal to 10}

\item{n.exact.t3}{: if \eqn{min\{m,n\}\leq n.exact} the critical values of the LMPI \eqn{T_3}
statistic are computed via permutation. Default value is set equal to 10}

\item{B.f}{: number of iteration to approximate Fisher test distribution via permutation. Default value is set equal to 10^3}

\item{B.t3}{: number of iteration to approximate \eqn{T_3} test distribution via permutation. Default value is set equal to 10^3}

\item{seed}{: seed to ensure reproducible results when approximating the test distribution via permutation}

\item{alpha}{: significance level. Default value is set equal to 0.1}
}
\value{
A vector of integer numbers corresponding to the \eqn{(1 − \alpha)}-confidence lower bound for the
number of true discoveries in closed testing procedure using the local tests chosen by the user.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
}
\description{
d_choose.method
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_choose.method(S_Y=Sy, S_X=Sx, methods = "T3")
d_choose.method(S_Y=Sy, S_X=Sx, methods = c("simes", "adjustedFisher", "t2", "T3"))


}
