% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_Fisher.corrected.R
\name{d_Fisher.corrected}
\alias{d_Fisher.corrected}
\title{d_Fisher.corrected}
\usage{
d_Fisher.corrected(S_Y, S_X, gamma = NULL, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{gamma}{: parameter related to the inflation of the variance of the Fisher test statistic when \emph{p}-values are not independent}

\item{alpha}{: significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
the corrected Fisher local test applied to conformal \emph{p}-values, that rejects
the global null hypothesis when
\deqn{\frac{-2\sum_{i=1}^n\log(p_i)+2n(\sqrt(1+\gamma)-1)}{\sqrt(1+\gamma)}\geq \chi^2_{2n}(1-\alpha),}
where \chi^2_{2n}(1-\alpha) is the \eqn{(1-\alpha)} quantile of the chi-squared
distribution with \eqn{2n} degree of freedom.
The selection set, i.e. the set of hypothesis
indices that we are interested in is \eqn{[n]=:\{1,...,n\}} by default.
}
\description{
d_Fisher.corrected
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_Fisher.corrected(S_Y=Sy, S_X=Sx, alpha=0.1)
}
