% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_G_fast.R
\name{d_G_monotone}
\alias{d_G_monotone}
\title{d_G_monotone}
\usage{
d_G_monotone(
  S_X,
  S_Y,
  S = NULL,
  g.hat,
  decr = FALSE,
  k = NULL,
  alpha = 0.1,
  pvalue_only = FALSE,
  n_perm = 0,
  B = 10^3,
  B_MC = 10^3,
  seed = 123
)
}
\arguments{
\item{S_X}{calibration score vector.}

\item{S_Y}{test score vector.}

\item{S}{selection set in the index test set. If \code{NULL} the entire test set is selected.}

\item{g.hat}{it can be either a character ("analytical") or a function denoting the outlier density.
If g.hat="analytical" the test statistics are computed analytically without Monte Carlo estimation.}

\item{decr}{logical value indicating whether the outlier distribution is decreasing (TRUE)
or increasing (FALSE).}

\item{k}{order of the LMPI test statistic to be specified when g.hat is "analytical".}

\item{alpha}{significance level.}

\item{pvalue_only}{logical value. If \code{TRUE}, only the global test is performed.}

\item{n_perm}{minimum test sample size needed to use the asymptotic distribution of the test statistic.}

\item{B}{number of replications to compute critical values and global \emph{p}-value. Default value is 10^3.}

\item{B_MC}{number of replications to compute the Shiraishi test statistic.}

\item{seed}{seed to ensure reproducible results.}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using the chosen local test.
\item \code{S}: a vector which is the selection set. If \code{NULL}, the selection set is the entire test set.
\item \code{global.pvalue}: a number which is the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null.
\item \code{selection.pvalue}: a number which is the \emph{p}-value for the selected null. By default it is set equal to 1.
}
}
\description{
It performs closed testing method with Shiraishi local test using an exact shortcut
valid when the outlier density is monotone, either increasing or decreasing.
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
rg2 = function(rnull, k=2) max(rnull(k))
m = 10; n=10;
X = runif(m)
Y = replicate(n, rg2(rnull=runif))
res = d_G_monotone(X, Y, S=c(1:7), g.hat=g2, decr=FALSE, B=100)
}
