% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitney.R
\name{d_MannWhitney}
\alias{d_MannWhitney}
\title{d_MannWhitney}
\usage{
d_MannWhitney(S_Y, S_X, n.exact = 10, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{n.exact}{: maximum value of the sample size of the second sample for which the critical
values of the Wilcoxon-Mann-Whitney statistic are exactly computed using \emph{\strong{qwilcox}} function.
Default value is set equal to 10.}

\item{alpha}{: significance level. Default level is set equal to 0.1}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Wilcoxon-Mann-Whitney local test that rejects the global null hypothesis at level \eqn{\alpha}
when the test statistic \deqn{\sum_{i=1}^n\sum_{j=1}^m \mathbb{1}\{X_j<Y_i\}} is greater than
the critical value corresponding to the significance level \eqn{\alpha}.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Wilcoxon-Mann-Whitney local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
crit = crit.WMW(m=length(Sx), n=length(Sy), alpha=0.1)
d_MannWhitney(S_Y=Sy, S_X=Sx, alpha=0.1)

}
