% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitneyk3.R
\name{d_MannWhitneyk3}
\alias{d_MannWhitneyk3}
\title{d_MannWhitneyk3}
\usage{
d_MannWhitneyk3(S_Y, S_X, alpha)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{alpha}{: significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
the LMPI local test corresponding to \eqn{k=3} applied to conformal \emph{p}-values.
The selection set, i.e. the set of hypothesis
indices that we are interested in is \eqn{[m]=:\{1,...,m\}} by default.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Wilcoxon-Mann-Whitney local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_MannWhitneyk3(S_Y=Sy, S_X=Sx, alpha=0.1)


}
