% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_G_fast.R
\name{d_selection_G}
\alias{d_selection_G}
\title{d_selection_G}
\usage{
d_selection_G(
  S_X,
  S_Y,
  S = NULL,
  k = NULL,
  g.hat = NULL,
  monotonicity = NULL,
  fit_method = "betamix",
  prop.F = 0.5,
  alpha = 0.1,
  pvalue_only = FALSE,
  n_perm = 10,
  B = 10^3,
  B_MC = 10^3,
  seed = 123
)
}
\arguments{
\item{S_X}{calibration score vector.}

\item{S_Y}{test score vector.}

\item{S}{selection set in the index test set. If \code{NULL} the entire test set is selected.}

\item{k}{order of the generalized Wilcoxon sum-rank test. Classic Wilcoxon sum-rank test corresponds to \eqn{k=1}.}

\item{g.hat}{it can be either a character ("analytical") or a function denoting the outlier density.
If g.hat="analytical" the test statistics are computed analytically without Monte Carlo estimation.
If \code{NULL}, the outlier density is estimated from the data.}

\item{monotonicity}{character indicating if the outlier density function is monotone increasing or decreasing or neither. Default value is \code{NULL}.}

\item{fit_method}{character value indicating the method to approximate the outlier distribution when argument \code{g.hat} is \code{NULL}.
It can be either "beta_mix" or "mixmodel".}

\item{prop.F}{proportion of inliers used to estimate the inlier distribution in the process of estimating the outlier density.
Default value is 0.5.}

\item{alpha}{significance level.}

\item{pvalue_only}{logical value. If \code{TRUE}, only the global test is performed.}

\item{n_perm}{minimum test sample size needed to use the asymptotic distribution of the test statistic}

\item{B}{number of replications to compute critical values and global \emph{p}-value. Default value is 10^3.}

\item{B_MC}{number of replications to compute the Shiraishi test statistic.}

\item{seed}{seed to ensure reproducible results.}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using the chosen local test.
\item \code{S}: a vector which is the selection set. If \code{NULL}, the selection set is the entire test set.
\item \code{global.pvalue}: a number which is the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null.
\item \code{selection.pvalue}: a number which is the \emph{p}-value for the selected null. By default it is set equal to 1.
}
}
\description{
It performs closed testing method with Shiraishi local test using an exact shortcut
valid when the outlier density is monotone, either increasing or decreasing.
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
rg2 = function(rnull, k=2) max(rnull(k))

X = runif(10)
Y = replicate(10, rg2(rnull=runif))
res = d_selection_G(X, Y, S = c(1:7), g.hat = g2, monotonicity="increasing", B=100)
}
