% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_simes.R
\name{d_selection_storey}
\alias{d_selection_storey}
\title{d_selection_storey}
\usage{
d_selection_storey(
  S_X,
  S_Y,
  S = NULL,
  alpha = 0.1,
  lambda = 0.5,
  pvalue_only = FALSE
)
}
\arguments{
\item{S_X}{: score vector of calibration observations.}

\item{S_Y}{: score vector of test observations.}

\item{S}{: selection set in the index test set. If \code{NULL} the entire test set is selected.}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1.}

\item{lambda}{: parameter involved in the computation of Storey estimator. Default value is set equal to 0.5.}

\item{pvalue_only}{: logical value. If \code{TRUE}, only the global test is performed.}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using the chosen local test.
\item \code{S}: a vector which is the selection set. If \code{NULL}, the selection set is the entire test set.
\item \code{global.pvalue}: a number which is the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null.
\item \code{selection.pvalue}: a number which is the \emph{p}-value for the selected null. By default it is set equal to 1.
}
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test with Storey estimator for the proportion of true null hypotheses.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_selection_storey(Sx, Sy)
d_selection_storey(Sx, Sy, S=3)
d_selection_storey(Sx, Sy, S=c(3, 7:13))

}
