% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_Simes.R
\name{d_simes}
\alias{d_simes}
\title{d_simes}
\usage{
d_simes(S_Y, S_X, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector of test observations}

\item{S_X}{: score vector of calibration observations}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1.}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for the
number of true discoveries in closed testing procedure using Simes local
test applied to conformal \emph{p}-values.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_simes(S_Y=Sy, S_X=Sx)


}
