% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_StoreyBH.R
\name{d_StoreyBH}
\alias{d_StoreyBH}
\title{d_StoreyBH}
\usage{
d_StoreyBH(S_Y, S_X, alpha = 0.1, lambda = 0.5)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{alpha}{: significance level. Default level is set equal to 0.1}

\item{lambda}{: parameter involved in the computation of Storey estimator. Default value is set equal to 0.5}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for the
number of true discoveries in Benjamini-Hochberg procedure with Storey estimator
for the proportion of true null hypotheses applied to conformal \emph{p}-values.
The selection set, i.e. the set of hypothesis indices that we are
interested in is \eqn{[m]=:\{1,...,m\}} by default.
Then, Storey estimator is computed as
\deqn{\hat\pi_0 = \frac{1+\sum_{i=1}^m \mathbb{1}\{p_i>\lambda\}}{m(1-\lambda)}}
where \eqn{\lambda\in(0,1)} and \eqn{p_i} is the \emph{p}-value related
to the null hypothesis \eqn{H_i, \hspace{2mm} i=1,\ldots,m}.
}
\description{
d_StoreyBH
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_StoreyBH(S_Y=Sy, S_X=Sx)



}
