% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t2.R
\name{d_t2}
\alias{d_t2}
\title{d_t2}
\usage{
d_t2(S_X, S_Y, n.exact = 10, alpha = 0.1)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{n.exact}{: if \eqn{min\{m,n\}\leq n.exact} the critical values of the Wilcoxon-Mann-Whitney
statistic are exactly computed using \emph{\strong{qwilcox}} function. Default value is set equal to 10}

\item{alpha}{: significance level. Default level is set equal to 0.1}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Wilcoxon-Mann-Whitney local test that rejects the global null hypothesis at level \eqn{\alpha}
when the test statistic \deqn{\sum_{i=1}^n\sum_{j=1}^m \mathbb{1}\{X_j<Y_i\}} is greater than
the critical value corresponding to the significance level \eqn{\alpha}.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Wilcoxon-Mann-Whitney local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_t2(S_Y=Sy, S_X=Sx, alpha=0.1)

}
