% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dselection.singleton_MannWhitney.R
\name{dselection.singleton_MannWhitney}
\alias{dselection.singleton_MannWhitney}
\title{dselection.singleton_MannWhitney}
\usage{
dselection.singleton_MannWhitney(S_Y, S_X, S, alpha)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{S}{: selected set of indices. Must be a singleton.}

\item{alpha}{: significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Wilcoxon-Mann-Whitney local test applied to conformal \emph{p}-values.
The selection set S, i.e. the set of hypothesis
indices that we are interested in, must be a singleton.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Wilcoxon-Mann-Whitney local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
crit = nout::crit.WMW(m=length(Sx), n=length(Sy), alpha=0.1)$crit.val
dselection.singleton_MannWhitney(S_Y=Sy, S_X=Sx, S=3, alpha=0.1)

}
