% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dselection_Simes.R
\name{dselection_Simes}
\alias{dselection_Simes}
\title{dselection_Simes}
\usage{
dselection_Simes(S_Y, S_X, selection, alpha = 0.1)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{selection}{: vector of selected indices}

\item{alpha}{: significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Simes local test applied to conformal \emph{p}-values. The selection set, i.e., the set of hypothesis
indices that we are interested can be choosen.
}
\description{
dselection_Simes
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
dselection_Simes(S_Y=Sy, S_X=Sx, selection = 5:15)

}
