% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dselection_StoreySimes.R
\name{dselection_StoreySimes}
\alias{dselection_StoreySimes}
\title{dselection_Simes}
\usage{
dselection_StoreySimes(S_X, S_Y, S, lambda = 0.5, alpha)
}
\arguments{
\item{S_X}{vector of calibration scores}

\item{S_Y}{vector of test scores}

\item{S}{set of indices corresponding to the \emph{p}-values to be tested}

\item{lambda}{: parameter involved in the computation of Storey estimator.
Default value is set equal to 0.5}

\item{alpha}{significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Simes local test with Storey's estimator for the proportion of true null
hypotheses applied to conformal \emph{p}-values. The selection set, i.e. the set of hypothesis
indices that we are interested in can be chosen.
}
\description{
dselection_Simes
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
dselection_StoreySimes(S_Y=Sy, S_X=Sx, S = 1:15, alpha = 0.2)

}
