% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t2.R
\name{exact.crit.WMW}
\alias{exact.crit.WMW}
\title{exact.crit.WMW}
\usage{
exact.crit.WMW(m, n, alpha = 0.1)
}
\arguments{
\item{m}{: number of observations in the calibration set}

\item{n}{: number of observations in the test set}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}
}
\value{
A R object of class \emph{crit.val.info}, which is a list consisting
of the following elements: \itemize{
\item m number of calibration observations
\item n number of test observations
\item alpha significance level of the local test
\item crit.val vector of critical values of the Wilcoxon-Mann-Whitney
statistic \eqn{U(n,k)}, \eqn{k=m,\ldots,1} (first element of the vector
corresponds to \eqn{k=m} and the last one to \eqn{k=1.} Given two samples
\eqn{(X_1,\ldots,X_n)} and \eqn{(Y_1,\ldots,Y_k)}, the Wilcoxon-Mann-Whitney
statistic is computed as \deqn{U(n,k) = \sum_{i = 1}^{n}\sum_{j = 1}^{k} \mathbb{1}\{X_i > Y_j\}}
and it has mean \eqn{kn/2} and variance \eqn{kn(k+n+1)/12}.}
}
\description{
Given \eqn{m} observations in the calibration set and \eqn{n} observations in the test set,
it relies on R function \emph{\strong{qwilcox}} and returns the vector of critical values of the
Wilcoxon-Mann-Whitney test statistic for each level of the closed testing.
}
\examples{

exact.crit.WMW(n=10,m=5,alpha=0.1)

n=8;m=10
crits = exact.crit.WMW(m=m, n=n,alpha=0.1)$crit.val
plot(x=1:n, y=crits, xlab="n-k+1", main = "Critical values of WMW for m=10 and k=1,...,n=8")



}
