% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findd.R
\name{findd}
\alias{findd}
\title{findd}
\usage{
findd(method, S_X, S_Y, S, lambda = 0.5, alpha)
}
\arguments{
\item{method}{a string indicating the method to be used.
Possible values are "Simes", "StoreySimes", "MannWhitney"}

\item{S_X}{values of calibration scores}

\item{S_Y}{values of test scores}

\item{S}{set of selected indices corresponding to the observations to be tested}

\item{lambda}{parameter involved in the computation of Storey estimator. Default value is set equal to 0.5}

\item{alpha}{significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for the
number of true discoveries in closed testing procedure using the selected method as local
test.
}
\description{
findd
}
\examples{

set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
findd(method = "MannWhitney", S=1:length(Sy), S_Y=Sy, S_X=Sx, alpha=0.1)
findd(method = "Simes", S=1:length(Sy), S_Y=Sy, S_X=Sx, alpha=0.1)
findd(method = "StoreySimes", S=1:10, S_Y=Sy, S_X=Sx, alpha=0.1)

}
