% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{fit_beta_mixture}
\alias{fit_beta_mixture}
\title{fit_beta_mixture}
\usage{
fit_beta_mixture(data, num_starts = 10)
}
\arguments{
\item{data}{score vector}

\item{num_starts}{number of iterations for parameter optimization}
}
\value{
A list of three numbers that are, respectively, the estimates of
the first parameter (alpha) of a Beta distribution, of the second parameter (beta) of a Beta distribution and
of the mixing proportion (lambda) of the mixture distribution.
}
\description{
It estimates the parameters of a mixture model of two distributions,
where the first one is a Standard Uniform and the second one is a Beta distribution
whose parameters are both unknown and also the mixing proportion is unknown.
All the parameters are optimized minimizing the negative log-likelihood,
allowing for box constraints.
}
