% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{inverse_transform_sampling}
\alias{inverse_transform_sampling}
\title{inverse_transform_sampling}
\usage{
inverse_transform_sampling(
  cdf_function,
  n_samples = 1000,
  lower_bound = 0,
  upper_bound = 1
)
}
\arguments{
\item{cdf_function}{cumulative density function.}

\item{n_samples}{number of samples to be generated.}

\item{lower_bound}{lower bound of the density support.}

\item{upper_bound}{upper bound of the density support.}
}
\value{
A vector of samples from the desired distribution of the prespecified length.
}
\description{
Given a cumulative density function, it generates a specific number of samples
from the specified distribution via inverse transform sampling.
}
