% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{make_density_monotone_decreasing}
\alias{make_density_monotone_decreasing}
\title{make_density_monotone_decreasing}
\usage{
make_density_monotone_decreasing(g)
}
\arguments{
\item{g}{a density function}
}
\value{
A monotone decreasing density function that is the output of Isotonic Regression
}
\description{
It makes a density function monotone decreasing using Isotonic Regression
}
