% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_adjFisher.R
\name{perm.crit.adjFisher}
\alias{perm.crit.adjFisher}
\title{perm.crit.adjFisher}
\usage{
perm.crit.adjFisher(S_X, S_Y, alpha = 0.1, B = 10^3, seed = 123)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}

\item{B}{: number of interations to approximate the adjusted Fisher test statistic distribution via permutation.
Default value is set equal to \eqn{10^3}.}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A R object of class \emph{crit.val.info}, which is a list consisting
of the following elements: \itemize{
\item m number of test observations
\item n number of calibration observations
\item alpha significance level of the local test
\item crit.val critical value of the adjusted Fisher test statistic.}
}
\description{
Given \eqn{m} observations in the calibration set and \eqn{n} observations in the test set,
it returns the vector of critical values at level \eqn{\alpha} of the adjusted Fisher test statistic
\deqn{\frac{-2\sum_{i=1}^n \log(p_i)+2n(\sqrt{1+\gamma}-1)}{\sqrt{1+\gamma}}}
where \eqn{\gamma\in(0,+\infty)}.
}
\examples{
crits = perm.crit.adjFisher(S_X=runif(10),S_Y=runif(min=0, max=0.5, 5),alpha=0.1)


}
