% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{perm.crit.T}
\alias{perm.crit.T}
\title{perm.crit.T}
\usage{
perm.crit.T(
  m,
  n,
  local_test = "wmw",
  k = NULL,
  stats_G_vector = NULL,
  alpha = 0.1,
  B = 10^3,
  seed = 123
)
}
\arguments{
\item{m}{calibration sample size.}

\item{n}{test sample size.}

\item{local_test}{local test to be used in the closed testing procedure. Default value is Wilcoxon sum-rank test.}

\item{k}{order of the higher order Wilcoxon sum-rank test statistic.}

\item{stats_G_vector}{vector of elementary test statistics to perform the test in Shiraishi (1985).}

\item{alpha}{significance level. Default value is set equal to 0.1.}

\item{B}{number of permutations.}

\item{seed}{seed to ensure reproducible results.}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of the chosen test statistic obtained via permutation.
}
\description{
It computes the permuted \eqn{(1-\alpha)}-quantile of a chosen test statistic.
}
