% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitneyk3.R
\name{perm.crit.T3}
\alias{perm.crit.T3}
\title{perm.crit.T3}
\usage{
perm.crit.T3(S_X, S_Y, alpha = 0.1, B = 10^3, seed = 123)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}

\item{B}{: number of interations to approximate \eqn{T_3} distribution via permutation.
Default value is set equal to \eqn{10^3}.}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A R object of class \emph{crit.val.info}, which is a list consisting
of the following elements: \itemize{
\item m number of calibration observations
\item n number of test observations
\item alpha significance level of the local test
\item crit.val critical value of LMPI \eqn{T_3} statistic for fixed values of \eqn{m,n} and \eqn{\alpha}.}
}
\description{
Given \eqn{m} observations in the calibration set and \eqn{n} observations in the test set,
it returns the vector of critical values at level \eqn{\alpha} of the LMPI \eqn{T_3} test statistic at level \eqn{\alpha}.
}
\examples{
perm.crit.T3(S_X=runif(10),S_Y=runif(min=0, max=0.5, 5),alpha=0.1)


}
