% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{standardize_to_uniform}
\alias{standardize_to_uniform}
\title{standardize_to_uniform}
\usage{
standardize_to_uniform(X1, X2)
}
\arguments{
\item{X1}{a vector}

\item{X2}{a vector}
}
\value{
A list of two vectors corresponding to the scaled ranks of the two input vectors
}
\description{
It rank the data and scale ranks to \verb{[0, 1]}.
}
