% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{stat.T2}
\alias{stat.T2}
\title{stat.T2}
\usage{
stat.T2(Z, m)
}
\arguments{
\item{Z}{: pooled score vector with the first \eqn{m} components corresponding
to the calibration observations and the last \eqn{n} components corresponding
to the calibration observations}

\item{m}{: calibration sample size}
}
\value{
Given the pooled score vector \eqn{Z=(X,Y)} where \eqn{X} is the calibration score
vector and \eqn{Y} is the test score vector, for each observation in the test sample
it returns its rank:\deqn{\sum{j=1}^m\mathbb{1}\{X_j<Y_i\}.}
}
\description{
It computes the ranks of the test observations in the pooled
vector of calibration and test scores.
}
