% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{stat.Tk}
\alias{stat.Tk}
\title{stat.Tk}
\usage{
stat.Tk(Z, m, k)
}
\arguments{
\item{Z}{pooled score vector with the first \eqn{m} components corresponding
to calibration observations and the last \eqn{n} components corresponding
to test observations.}

\item{m}{calibration sample size.}

\item{k}{order of the LMPI test statistic.}
}
\value{
Given the pooled score vector \eqn{Z=(X,Y)} where \eqn{X} is the calibration score
vector and \eqn{Y} is the test score vector, for each observation in the test sample
it returns
\deqn{T_{k,i}=\sum{j=1}^m\mathbb{1}\{X_j<Y_i\}+\left(\sum{j=1}^m\mathbb{1}\{X_j<Y_i\}\right)^2+\ldots+\left(\sum{j=1}^m\mathbb{1}\{X_j<Y_i\}\right)^k.}
}
\description{
It computes the ranks of the test observations in the pooled
vector of calibration and test scores. Then, component-wisely compute the power
from \eqn{1} to \eqn{k} of the rank vector and component-wisely sum them.
}
