% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fisher.R
\name{asymptotic.critical.Fisher}
\alias{asymptotic.critical.Fisher}
\title{asymptotic.critical.Fisher}
\usage{
asymptotic.critical.Fisher(m, n, alpha = 0.1)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{alpha}{: significance level. Default value is set equal to 0.1}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of adjusted Fisher test statistic
based on asymptotic Chi-squared approximation
\deqn{\frac{-2\sum_{i=1}^n\log(p_i) + 2n\sqrt{\gamma+1}-1}{\sqrt{\gamma+1}} \overset{d}{\sim} \chi^2_{2n},}
where \eqn{\gamma\in(0,+\infty)} is chosen equal to \eqn{n/m}, \eqn{p_i} are the conformal \emph{p}-values and
\eqn{\chi^2_{2n}} is the Chi-squared distribution with \eqn{2n} degrees of freedom.
}
\description{
It computes the \eqn{(1-\alpha)}-quantile of the adjusted Fisher test statistic
based on asymptotic Chi-squared approximation.
}
