% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.critical.Tk}
\alias{asymptotic.critical.Tk}
\title{asymptotic.critical.Tk}
\usage{
asymptotic.critical.Tk(m, n, k, alpha = 0.1)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}

\item{alpha}{: significance level. Default value is set equal to 0.1}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_k} test statistic
based on asymptotic normal approximation.
}
\description{
It computes the \eqn{(1-\alpha)}-quantile of LMPI \eqn{T_k} test statistic
based on asymptotic normal approximation.
}
