% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{asymptotic.moments.Tk}
\alias{asymptotic.moments.Tk}
\title{asymptotic.moments.Tk}
\usage{
asymptotic.moments.Tk(m, n, k)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}
}
\value{
It returns the mean and the variance of the asymptotic distribution of \eqn{T_k}.
For \eqn{k=1,2,3} the exact mean is given, i.e., the finite-sample mean,
and the provided estimate of the variance is more accurate than the first order asymptotic approximation.
}
\description{
It computes the mean and the variance of the asymptotic distribution of \eqn{T_k}.
For \eqn{k=1,2,3} the exact mean is given, i.e., the finite-sample mean,
and the provided estimate of the variance is more accurate than the first order asymptotic approximation.
}
