% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{asymptotic.moments.Tk}
\alias{asymptotic.moments.Tk}
\title{asymptotic.moments.Tk}
\usage{
asymptotic.moments.Tk(m, n, k)
}
\arguments{
\item{m}{: calibration sample size.}

\item{n}{: test sample size.}

\item{k}{: order of the LMPI test statistic.}
}
\value{
A list with the mean and the variance of the asymptotic distribution of the higher order Wilcoxon sum-rank test statistic.
}
\description{
It computes the mean and the variance of the asymptotic distribution of the higher order Wilcoxon sum-rank test statistic.
}
