% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.pvalue.Fisher}
\alias{asymptotic.pvalue.Fisher}
\title{asymptotic.pvalue.Fisher}
\usage{
asymptotic.pvalue.Fisher(m, n, T.obs)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{T.obs}{: observed value of the test statistic}
}
\value{
It returns the approximated \emph{p}-value of the adjusted Fisher
test statistic based on the asymptotic Chi-squared approximation
\deqn{\frac{-2\sum_{i=1}^n\log(p_i) + 2n\sqrt{\gamma+1}-1}{\sqrt{\gamma+1}} \overset{d}{\sim} \chi^2_{2n},}
where \eqn{\gamma\in(0,+\infty)} is chosen equal to \eqn{n/m}, \eqn{p_i} are the conformal \emph{p}-values and
\eqn{\chi^2_{2n}} is the Chi-squared distribution with \eqn{2n} degrees of freedom.
}
\description{
It computes the approximated \emph{p}-value of the adjusted Fisher
test statistic based on the asymptotic Chi-squared approximation.
}
