% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{asymptotic.pvalue.Tk}
\alias{asymptotic.pvalue.Tk}
\title{asymptotic.pvalue.Tk}
\usage{
asymptotic.pvalue.Tk(m, n, k, T.obs)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}

\item{T.obs}{: observed value of the test statistic}
}
\value{
It returns the approximated \emph{p}-value of the LMPI \eqn{T_k}
test statistic based on the asymptotic normal approximation
}
\description{
It computes the approximated \emph{p}-value of the LMPI \eqn{T_k}
test statistic based on the asymptotic normal approximation.
}
