% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{beta_moment.k}
\alias{beta_moment.k}
\title{beta_moment.k}
\usage{
beta_moment.k(h, N, k)
}
\arguments{
\item{h}{: an integer between \eqn{1} and \eqn{N}}

\item{N}{: integer}

\item{k}{: order of moment}
}
\value{
\eqn{k}the moment of a Beta(h, N-h+1) distribution
}
\description{
beta_moment.k
}
\examples{
mom = beta_moment.k(h=5,N=10,k=3)
}
