% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{choose_best_monotonic_density}
\alias{choose_best_monotonic_density}
\title{choose_best_monotonic_density}
\usage{
choose_best_monotonic_density(g)
}
\arguments{
\item{g}{: a density function}
}
\value{
A density function that is the output of Isotonic Regression and
is monotone decreasing or increasing based on the lower Residual Sum of Squares
}
\description{
It chooses between increasing and decreasing monotonicity
}
