% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{compute.1critical.value}
\alias{compute.1critical.value}
\title{compute.1critical.value}
\usage{
compute.1critical.value(
  m,
  n,
  local_test = "wmw",
  alpha,
  k = NULL,
  stats_G_vector = NULL,
  n_perm = 0,
  B = 10^3,
  critical_values = NULL,
  seed = 123
)
}
\arguments{
\item{m}{: calibration size}

\item{n}{: test size}

\item{local_test}{: local test to be used in the closed testing procedure. Default value is Wilcoxon test}

\item{alpha}{: significance level}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}

\item{n_perm}{: if \eqn{min(m,n)\leq n_perm} critical values will be computed via permutation. Default value is 0}

\item{B}{: number of permutation to compute critical values. Default value is 10^3}

\item{critical_values}{: if not \code{NULL}, a vector of precomputed critical values obtained using
the permutation distribution of the test statistic}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A ritical value for a test statistic chosen among LMPI \eqn{T_k}, Fisher or Shiraishi statistics
at significance level \eqn{\alpha} with calibration size \eqn{m} fixed for each level of closed testing.
}
\description{
It computes the critical value for a chosen test statistic
at significance level \eqn{\alpha}.
}
