% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{compute.critical.values}
\alias{compute.critical.values}
\title{compute.critical.values}
\usage{
compute.critical.values(
  m,
  n,
  alpha,
  k = NULL,
  correlation_remainder,
  n_perm = 10,
  B = 10^3,
  critical_values = NULL,
  seed = 123
)
}
\arguments{
\item{m}{: calibration size}

\item{n}{: test size}

\item{alpha}{: significance level}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{correlation_remainder}{: correlation between the remainder and the $U$-statistic when statistic $T_3$ is used}

\item{n_perm}{: if \eqn{min(m,n)\leq n_perm} critical values will be computed via permutation. Default value is 10}

\item{B}{: number of permutation to compute critical values. Default value is 10^3}

\item{critical_values}{: if not \code{NULL}, a vector of precomputed critical values obtained using
the permutation distribution of the test statistic}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A vector of critical values for a test statistic chosen among LMPI \eqn{T_k} or Fisher statistics
at significance level \eqn{\alpha} with calibration size \eqn{m} fixed for each level of closed testing.
}
\description{
It computes the vector of critical values for a chosen test statistic
at significance level \eqn{\alpha}.
}
