% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{compute.critical.values}
\alias{compute.critical.values}
\title{compute.critical.values}
\usage{
compute.critical.values(
  m,
  n,
  alpha,
  stat.func,
  asymptotic.critical.func,
  n_perm = 10,
  B = 10^3,
  critical_values = NULL,
  seed = 123
)
}
\arguments{
\item{m}{: calibration size}

\item{n}{: test size}

\item{alpha}{: significance level}

\item{stat.func}{: test statistic of which compute critical values}

\item{asymptotic.critical.func}{: asymptotic approximation of \code{stat.func}}

\item{n_perm}{: if \eqn{min(m,n)\leq n_perm} critical values will be computed via permutation. Default value is 10}

\item{B}{: number of permutation to compute critical values. Default value is 10^3}

\item{critical_values}{: if not \code{NULL}, a vector of precomputed critical values obtained using
the permutation distribution of the test statistic}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A vector of critical values for a test statistic chosen among \eqn{T_2, T_3} or Fisher
at significance level \eqn{\alpha} with calibration size \eqn{m} fixed for each level of closed testing.
}
\description{
compute.critical.values
}
