% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{compute.global.pvalue}
\alias{compute.global.pvalue}
\title{compute.global.pvalue}
\usage{
compute.global.pvalue(
  T.obs,
  m,
  n,
  k = NULL,
  correlation_remainder,
  n_perm = 10,
  B = 100,
  seed = 321
)
}
\arguments{
\item{T.obs}{: observed value of the chosen test statistic}

\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{correlation_remainder}{: correlation between the remainder and the $U$-statistic when statistic $T_3$ is used}

\item{n_perm}{: if \eqn{min(m,n)\leq n_perm} the \emph{p}-value for the global null will be computed via permutation. Default value is 10}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \emph{p}-value for the global null, according to the chosen test statistic.
The \emph{p}-value is computed via permutation if either the calibration sample size or the test sample size
is smaller than \code{n_perm}. Otherwise, it is computed using the asymptotic distribution.
}
\description{
It computes the \emph{p}-value for the global null according to the chosen test statistic
}
