% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{compute.global.pvalue}
\alias{compute.global.pvalue}
\title{compute.global.pvalue}
\usage{
compute.global.pvalue(
  T.obs,
  m,
  n,
  local.test = "wmw",
  k = NULL,
  stats_G_vector = NULL,
  n_perm = 10,
  B = 100,
  seed = 321
)
}
\arguments{
\item{T.obs}{: observed value of the chosen test statistic}

\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{local.test}{: local test to be used in the closed testing procedure. Default value is Wilcoxon test}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}

\item{n_perm}{: if \eqn{min(m,n)\leq n_perm} the \emph{p}-value for the global null will be computed via permutation. Default value is 10}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \emph{p}-value for the global null, according to the chosen test statistic.
The \emph{p}-value is computed via permutation if either the calibration sample size or the test sample size
is smaller than \code{n_perm}. Otherwise, it is computed using the asymptotic distribution.
}
\description{
It computes the \emph{p}-value for the global null according to the chosen test statistic
}
