% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{compute.global.pvalue.shirashi.adaptive}
\alias{compute.global.pvalue.shirashi.adaptive}
\title{compute.global.pvalue.shirashi}
\usage{
compute.global.pvalue.shirashi.adaptive(
  S_X,
  S_Y,
  prop_cal = 0.5,
  num_mc = 1000,
  fit_method = "betamix",
  monotone = FALSE
)
}
\arguments{
\item{S_X}{: calibration score vector.}

\item{S_Y}{: test score vector.}

\item{prop_cal}{: proportion of inlier observations used for estimating the inlier distribution.}

\item{num_mc}{: number of Monte Carlo iterations.}

\item{fit_method}{: character value indicating the method for estimating the PDF.
It can be either "betamix" or "mixmodel".}

\item{monotone}{: logical value. If \code{TRUE} the estimated PDF via the specified method
is made monotone. The direction (increasing or decreasing) is automatically chosen from the data.}
}
\value{
A number, the global \emph{p}-value for the Shiraishi test when the outlier density is estimated from the data.
}
\description{
Given the calibration and test score vectors, it computes the global \emph{p}-value
for the Shiraishi test after estimating the outlier density function.
}
