% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{compute.perm.pval}
\alias{compute.perm.pval}
\title{compute.perm.pval}
\usage{
compute.perm.pval(T.obs, m, n, k = NULL, B = 10^3, seed = 123)
}
\arguments{
\item{T.obs}{: observed value of the chosen test statistic}

\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \emph{p}-value for the global null obtained via permutation using the chosen test statistic.
}
\description{
It computes the \emph{p}-value for the global null obtained via permutation using a chosen test statistic.
}
