% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{compute.perm.pval}
\alias{compute.perm.pval}
\title{compute.perm.pval}
\usage{
compute.perm.pval(
  T.obs,
  local.test = "wmw",
  m,
  n,
  k = NULL,
  stats_G_vector = NULL,
  B = 10^3,
  seed = 123
)
}
\arguments{
\item{T.obs}{: observed value of the chosen test statistic}

\item{local.test}{: local test to be used in the closed testing procedure. Default value is Wilcoxon test}

\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \emph{p}-value for the global null obtained via permutation using the chosen test statistic.
}
\description{
It computes the \emph{p}-value for the global null obtained via permutation using a chosen test statistic.
}
