% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_variance.Tk}
\alias{compute_variance.Tk}
\title{compute_variance.Tk}
\usage{
compute_variance.Tk(m, n, k)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic}
}
\value{
It returns the variance of LMPI \eqn{T_k} in the limit,
computed according to the Central Limit Theorem for \eqn{U}-statistics.
}
\description{
compute_variance.Tk
}
