% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{create_cdf_interpolated}
\alias{create_cdf_interpolated}
\title{create_cdf_interpolated}
\usage{
create_cdf_interpolated(
  pdf_func,
  lower_bound = 0,
  upper_bound = 1,
  grid_points = 1000
)
}
\arguments{
\item{pdf_func}{: density function}

\item{lower_bound}{: lower bound of the density support}

\item{upper_bound}{: upper bound of the density support}

\item{grid_points}{: number of points where to compute the cumulative density function via integration of the density function}
}
\value{
A function which is the interpolated cumulative density function
}
\description{
It creates cumulative density function using linear interpolation within a grid of points
}
