% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitney.R
\name{critWMW}
\alias{critWMW}
\title{critWMW}
\usage{
critWMW(m, n, alpha, n.exact = 10, exact = F)
}
\arguments{
\item{m}{number of observations in the calibration set}

\item{n}{number of observations in the test set}

\item{alpha}{significance level of the local test. Default value is set equal to 0.1}

\item{n.exact}{maximum value of the sample size of the second sample for which the critical
values of the Wilcoxon-Mann-Whitney statistic are exactly computed using \emph{\strong{qwilcox}} function.
Default value is set equal to 10.}

\item{exact}{logical value. If TRUE, exact computation of critical values is performed for
all values of \eqn{k\in{1,\ldots,m}} when \eqn{n\leq 20}. Otherwise, exact computation of
critical values is performed for all values of \eqn{k\in\{1,\ldots,m.exact\}} when \eqn{n<1000}.
Default value is FALSE.}
}
\value{
A R object of class \emph{crit.vals.info}, which is a list consisting
of the following elements: \itemize{
\item m number of test observations
\item n number of calibration observations
\item alpha significance level of the local test
\item crit.vals vector of critical values of the Wilcoxon-Mann-Whitney
statistic \eqn{U(n,k)}, \eqn{k=m,\ldots,1} (first element of the vector
corresponds to \eqn{k=m} and the last one to \eqn{k=1.} Given two samples
\eqn{(X_1,\ldots,X_n)} and \eqn{(Y_1,\ldots,Y_k)}, the Wilcoxon-Mann-Whitney
statistic is computed as \deqn{U(n,k) = \sum_{i = 1}^{n}\sum_{j = 1}^{k} \mathbb{1}\{X_i > Y_j\}}
and it has mean \eqn{kn/2} and variance \eqn{kn(k+n+1)/12}.

When \eqn{n<1000} and \eqn{m<10} the value of the Wilcoxon-Mann-Whitney statistic
is exactly computed with the R function \emph{\strong{qwilcox}}; otherwise, we use
the normal approximation using continuity correction \deqn{U(n,k) \sim N(kn/2-1/2, kn(k+n+1)/12)}.
}
}
\description{
Given the number of observations in the calibration set (\eqn{n}) and the number
of observations in the test set (\eqn{m}), it returns the vector of critical
values \eqn{U(n,k)} of the Wilcoxon-Mann-Whitney test statistic letting \eqn{k} vary between
\eqn{1} and \eqn{m}, while the size of the first sample is kept fixed equal to \eqn{n}.
}
\examples{
critWMW(n=100,m=2000,alpha=0.1)

critWMW(n=100,m=900,alpha=0.1)

n=10;m=20
crits = critWMW(m=m, n=n,alpha=0.1)$crit.vals
plot(x=1:n, y=crits, xlab="n-k+1", main = "Critical values of WMW statistic for m=20 and k=1,...,n")



}
