% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_G.R
\name{d_G_monotone}
\alias{d_G_monotone}
\title{d_G_monotone}
\usage{
d_G_monotone(
  S_X,
  S_Y,
  S = NULL,
  stats_G_vector,
  alpha = 0.1,
  n_perm = 10,
  B = 10^3,
  seed = 123
)
}
\arguments{
\item{S_X}{:  calibration score vector}

\item{S_Y}{: test score vector}

\item{S}{: selection set in the index test set}

\item{stats_G_vector}{: list of Shiraishi (1985) test statistics for each closed testing level}

\item{alpha}{: significance level}

\item{n_perm}{: minimum test sample size needed to use the asymptotic distribution of the test statistic when
local.test is either "higher" or "fisher"}

\item{B}{: number of replications to compute critical values and global \emph{p}-value. Default value is 10^3}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using the chosen local test
\item \code{S}: the selection set, i.e., the selected subset of the test indices
\item \code{global.pvalue}: the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null
\item \code{selection.pvalue}: \emph{p}-value for the selected null
}
}
\description{
d_G_monotone
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
rg2 = function(rnull, k=2) max(rnull(k))
m = 50; n=50;
X = runif(m)
Y = replicate(n, rg2(rnull=runif))
stats_G = sapply(length(Y):1, function(h) stats_G_j_MC(N=m+h, g=g2, B=300))
res = d_G_monotone(S_Y=Y, S_X=X, S=c(1:40), stats_G_vector=stats_G, B=100)
}
