% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_WMW.R
\name{d_mannwhitney}
\alias{d_mannwhitney}
\title{d_mannwhitney}
\usage{
d_mannwhitney(S_Y, S_X, crit)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{crit}{: a R object of class \emph{crit.vals.info}}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
Wilcoxon-Mann-Whitney local test. The selection set, i.e. the set of hypothesis
indices that we are interested in is \eqn{[m]=:\{1,...,m\}} by default.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Wilcoxon-Mann-Whitney local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
crit = critWMW(m=length(Sy), n=length(Sx))
d_mannwhitney(S_Y=Sy, S_X=Sx)

}
