% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_MannWhitneyk3.R
\name{d_MannWhitneyk3}
\alias{d_MannWhitneyk3}
\title{d_MannWhitneyk3}
\usage{
d_MannWhitneyk3(S_Y, S_X, alpha)
}
\arguments{
\item{S_Y}{: score vector for the test set}

\item{S_X}{: score vector for the calibration set}

\item{alpha}{: significance level}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using
the LMPI local test corresponding to \eqn{k=3} applied to conformal \emph{p}-values.
The selection set is the set of indices corresponding to the observations in the test set that
we want to test. By default, the selection set is \eqn{[n]=:\{1,...,n\}} and corresponds to the
entire test set.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using the test statistic \eqn{\sum_{j=1}^n U_j(U_j+1)}, where \eqn{U_j} is the rank of the
\eqn{j}th observation in the test set among the vector of calibration units and \eqn{n} is the
number of observations in the test set of the test set.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_MannWhitneyk3(S_Y=Sy, S_X=Sx, alpha=0.1)




}
