% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_simes.R
\name{d_selection_simes}
\alias{d_selection_simes}
\title{d_selection_simes}
\usage{
d_selection_simes(S_Y, S_X, S = NULL, alpha = 0.1, pvalue_only = FALSE)
}
\arguments{
\item{S_Y}{: score vector of test observations}

\item{S_X}{: score vector of calibration observations}

\item{S}{: set of selected indices in the test set on which compute the lower bound for the number of outliers.
If S is missing, dthe global null hypothesis is used}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1.}

\item{pvalue_only}{: logical value. If TRUE, only the global test is performed}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in the selected set \eqn{S} in closed testing procedure using Simes' local test
\item \code{S}: the selection set, i.e., the selected subset of the test indices
\item \code{global.pvalue}: the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null
}
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_selection_simes(S_Y=Sy, S_X=Sx, S=3)
d_selection_simes(S_Y=Sy, S_X=Sx, S=c(3, 7:13))
d_selection_simes(S_Y=Sy, S_X=Sx)

}
