% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_selection_simes.R
\name{d_selection_storey}
\alias{d_selection_storey}
\title{d_selection_storey}
\usage{
d_selection_storey(
  S_Y,
  S_X,
  S = NULL,
  alpha = 0.1,
  lambda = 0.5,
  pvalue_only = FALSE
)
}
\arguments{
\item{S_Y}{: score vector of test observations}

\item{S_X}{: score vector of calibration observations}

\item{S}{: set of selected indices in the test set on which compute the lower bound for the number of outliers.
If S is missing, dthe global null hypothesis is used}

\item{alpha}{: significance level of the local test. Default value is set equal to 0.1}

\item{lambda}{: parameter involved in the computation of Storey estimator. Default value is set equal to 0.5}

\item{pvalue_only}{: logical value. If TRUE, only the global test is performed}
}
\value{
An integer which is the \eqn{(1 − \alpha)}-confidence lower bound for the
number of true discoveries in closed testing procedure using Simes local
test with Storey's estimator for the proportion of true null hypotheses applied to conformal \emph{p}-values.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
Then, Storey estimator is computed as
\deqn{\hat\pi_0 = \frac{1+\sum_{i=1}^n \mathbb{1}\{p_i>\lambda\}}{n(1-\lambda)}}
where \eqn{n} is the test sample size, \eqn{\lambda\in(0,1)} is a tuning parameter
and \eqn{p_i} is the \emph{p}-value corresponding to the \eqn{i}th hypothesis in the test set.
}
\description{
It returns the lower bound for the number of true discoveries in closed testing procedure
using Simes local test with Storey estimator for the proportion of true null hypotheses.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
d_selection_storey(S_Y=Sy, S_X=Sx)
d_selection_storey(S_Y=Sy, S_X=Sx, S=3)
d_selection_storey(S_Y=Sy, S_X=Sx, S=c(3, 7:13))
d_selection_storey(S_Y=Sy, S_X=Sx)



}
