% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveries_BH.R
\name{discoveries_BH}
\alias{discoveries_BH}
\title{discoveries_BH}
\usage{
discoveries_BH(S_X, S_Y, alpha = 0.1)
}
\arguments{
\item{S_X}{: score vector for the calibration set}

\item{S_Y}{: score vector for the test set}

\item{alpha}{: significance level. Default level is set equal to 0.1}
}
\value{
A vector corresponding to indices of test observations which are rejected
by Benjamini-Hochberg procedure applied to conformal \emph{p}-values.
The selection set is trivial, i.e., we are interested in testing all the observations in the test set by default.
}
\description{
Applies Benjamini-Hochberg procedure to conformal \emph{p}-values.
}
\examples{
set.seed(321)
Sxy = sample(x=1:1000, size=100)
Sx = sample(Sxy, size=70)
Sy = setdiff(Sxy, Sx)
discoveries_BH(Sx, Sy)


}
