% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_d.R
\name{find_d}
\alias{find_d}
\title{find_d}
\usage{
find_d(
  X,
  Y,
  local.test = "wmw",
  S = NULL,
  k = NULL,
  monotonicity = NULL,
  g.hat = NULL,
  alpha = 0.1,
  prop.F = 0.5,
  lambda = 0.5,
  n_perm = 0,
  B = 10^3,
  B_MC = 10^3,
  critical_values = NULL,
  seed = 123
)
}
\arguments{
\item{X}{: calibration score vector}

\item{Y}{: test score vector}

\item{local.test}{: local test to be used in the closed testing procedure.
It can be either "wmw" for Wilcoxon rank sum test, "higher" for higher order Wilcoxon rank sum tests,
"fisher" for Fisher's combination test, "g" for the test by Shiraishi (1985), "simes" for Simes' test or
"storey" for Simes' test using Storey's estimator for the proportion of true null hypotheses.}

\item{S}{: selection set in the index test set}

\item{k}{: positive integer indicating the order of generalized Wilcoxon rank sum test. Default value is NULL}

\item{monotonicity}{: logical value indicating if the outlier density function is monotone increasing or decreasing or neither.
It must be specified when the local test is "g"}

\item{g.hat}{: it denotes the outlier density. If NULL it is estimated from the data}

\item{alpha}{: significance level}

\item{prop.F}{: proportion of inliers used to estimate the inliser distribution while estimating the outlier density.
Default value is 0.5}

\item{lambda}{: parameter to be specified when computing Storey's estimator. Default value is 0.5}

\item{n_perm}{: minimum test sample size needed to use the asymptotic distribution of the test statistic when
local.test is either "higher" or "fisher"}

\item{B}{: number of replications to compute critical values and global \emph{p}-value. Default value is 10^3}

\item{B_MC}{: number of replications to compute the Shiraishi test statistic}

\item{critical_values}{: if not \code{NULL}, a vector of precomputed critical values obtained using
the permutation distribution of the test statistic}

\item{seed}{: seed to ensure reproducible results}
}
\value{
A list:
\itemize{
\item \code{lower_bound}: an integer which is the \eqn{(1 − \alpha)}-confidence lower bound for
the number of true discoveries in closed testing procedure using the chosen local test
\item \code{S}: the selection set, i.e., the selected subset of the test indices
\item \code{global.pvalue}: the global \emph{p}-value, i.e., the \emph{p}-value that closed testing procedure uses to reject the global null
\item \code{selection.pvalue}: \emph{p}-value for the selected null
}
}
\description{
find_d
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
rg2 = function(rnull, k=2) max(rnull(k))

X = runif(50)
Y = replicate(50, rg2(rnull=runif))
res = find_d(X, Y, B=100)
res = find_d(X, Y, local.test="higher", k=3, B=100)
res = find_d(X, Y, local.test="g", g.hat = g2, monotonicity=TRUE, B=100)
}
