% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{fit_mixmodel}
\alias{fit_mixmodel}
\title{fit_mixmodel}
\usage{
fit_mixmodel(data)
}
\arguments{
\item{data}{: vector of univariate scores}
}
\value{
A function which the PDF estimated via kernel smoothing from the score vector in input
}
\description{
It estimates the CDF from a vector of unidimensional scores and then, from the estimated CDF, it returns the PDF estimated via kernel smoothing
}
