% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{k_mom_beta}
\alias{k_mom_beta}
\title{k_mom_beta}
\usage{
k_mom_beta(a, b, k)
}
\arguments{
\item{a}{: first parameter of Beta distribution}

\item{b}{: second parameter of Beta distribution}

\item{k}{: order of the moment to be computed}
}
\value{
A real number
}
\description{
This function computes the \eqn{k}th moment of a Beta(a,b)
}
