% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{kernel_smoothed_pdf_function}
\alias{kernel_smoothed_pdf_function}
\title{kernel_smoothed_pdf_function}
\usage{
kernel_smoothed_pdf_function(cdf_function, n_samples = 1000)
}
\arguments{
\item{cdf_function}{: a cumulative density function}

\item{n_samples}{: number of samples to be generated to estimate the density function via Kernel Density Estimation}
}
\value{
A function which is the normalized Kernel Density Estimate of the density function
}
\description{
Given the cumulative density function, it returns a normalized Kernel Density Estimate of the density function
}
