% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{make_density_monotone_increasing}
\alias{make_density_monotone_increasing}
\title{make_density_monotone_increasing}
\usage{
make_density_monotone_increasing(g)
}
\arguments{
\item{g}{: a density function}
}
\value{
A monotone increasing density function that is the output of Isotonic Regression
}
\description{
It makes a density function monotone increasing using Isotonic Regression
}
