% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{mean_analytical.Lk}
\alias{mean_analytical.Lk}
\title{mean_analytical.Lk}
\usage{
mean_analytical.Lk(N, k, n)
}
\arguments{
\item{N}{: pooled sample size}

\item{k}{: an integer greater than 1, order of the Lehmann's alternative distribution}

\item{n}{: test sample size}
}
\value{
Asymptotic mean of Shiraishi (1985) test statistic
}
\description{
mean_analytical.Lk
}
\examples{
m = mean_analytical.Lk(N=100,k=2,n=50)
}
