% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{mean_analytical.Lk}
\alias{mean_analytical.Lk}
\title{mean_analytical.Lk}
\usage{
mean_analytical.Lk(N, k, n)
}
\arguments{
\item{N}{: pooled sample size.}

\item{k}{: an integer which is the order of the Lehmann's alternative distribution.}

\item{n}{: test sample size.}
}
\value{
A number which is the asymptotic mean of Shiraishi test statistic under Lehmann's alternative of order k.
}
\description{
It compute the analytical (not estimated through Monte Carlo simulation)
asymptotic mean of the Shiraishi test statistic.
}
