% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{meanG}
\alias{meanG}
\title{meanG}
\usage{
meanG(n, stats_G_vector)
}
\arguments{
\item{n}{: test size}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}
}
\value{
A numeric value corresponding to the mean of the asymptotic distribution of Shiraishi test (1985)
}
\description{
meanG
}
\examples{
n = 500
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
stats_G = stats_G_j_MC(N=1000, g=g2, B=10^3)
mu = meanG(n=n, stats_G_vector = stats_G)

}
