% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g_old.R
\name{meanG}
\alias{meanG}
\title{meanG}
\usage{
meanG(n, stats_G_vector)
}
\arguments{
\item{n}{: test size.}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985).}
}
\value{
A numeric value corresponding to the mean of the asymptotic distribution of Shiraishi test (1985).
}
\description{
meanG
}
