% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{perm.crit.T}
\alias{perm.crit.T}
\title{perm.crit.T}
\usage{
perm.crit.T(m, n, k = NULL, alpha = 0.1, B = 10^3, seed = 123)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{alpha}{: significance level. Default value is set equal to 0.1}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of the chosen test statistic obtained via permutation.
}
\description{
It computes permutation \eqn{(1-\alpha)}-quantile of a chosen test statistic.
}
