% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global_functions.R
\name{perm.crit.T}
\alias{perm.crit.T}
\title{perm.crit.T}
\usage{
perm.crit.T(
  m,
  n,
  local.test = "wmw",
  k = NULL,
  stats_G_vector = NULL,
  alpha = 0.1,
  B = 10^3,
  seed = 123
)
}
\arguments{
\item{m}{: calibration sample size}

\item{n}{: test sample size}

\item{local.test}{: local test to be used in the closed testing procedure. Default value is Wilcoxon test}

\item{k}{: order of the LMPI test statistic. If \code{NULL} it refers to Fisher test statistic}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}

\item{alpha}{: significance level. Default value is set equal to 0.1}

\item{B}{: number of permutations}

\item{seed}{: seed to ensure reproducible results}
}
\value{
It returns the \eqn{(1-\alpha)}-quantile of the chosen test statistic obtained via permutation.
}
\description{
It computes permutation \eqn{(1-\alpha)}-quantile of a chosen test statistic.
}
