% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_t.R
\name{stat.Fisher}
\alias{stat.Fisher}
\title{stat.Fisher}
\usage{
stat.Fisher(Z, m)
}
\arguments{
\item{Z}{: pooled score vector with the first \eqn{m} components corresponding
to the calibration observations and the last \eqn{n} components corresponding
to the calibration observations}

\item{m}{: calibration sample size}
}
\value{
Given the pooled score vector \eqn{Z=(X,Y)} where \eqn{X} is the calibration score
vector and \eqn{Y} is the test score vector, for each observation in the test sample
it returns the following quantity: \deqn{-log(p_i),}
where \eqn{p_i = \frac{1+\sum{j=1}^m\mathbb{1}\{X_j<Y_i\}}{m+1}} is the conformal
\emph{p}-value of the \eqn{i}th observation in the test set.
}
\description{
It computes the inverse of conformal \emph{p}-values for each observations in the test set.
Then, they are transformed applying the logarithmic function.
}
