% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{stat.G}
\alias{stat.G}
\title{stat.G}
\usage{
stat.G(Z, m, stats_G_vector)
}
\arguments{
\item{Z}{: pooled score vector}

\item{m}{: calibration size}

\item{stats_G_vector}{: vector of Shiraishi (1985) test statistics for a fixed test sample size}
}
\value{
vector of Shiraishi test statistics given a pooled score vector
}
\description{
stat.G
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
rg2 = function(rnull, k=2) max(rnull(k))
stats_G = stats_G_j_MC(N=1000, g=g2, B=500)
stat.G = function(Z=c(runif(50),replicate(rg2(runif), 50)), m=50,stats_G_vector=stats_G)

}
