% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g_old.R
\name{stat.G}
\alias{stat.G}
\title{stat.G}
\usage{
stat.G(Z, m, stats_G_vector)
}
\arguments{
\item{Z}{: pooled score vector.}

\item{m}{: calibration size.}

\item{stats_G_vector}{: vector of Shiraishi (1985) test statistics for a fixed test sample size.}
}
\value{
vector of Shiraishi test statistics given a pooled score vector.
}
\description{
stat.G
}
