% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{stat.MW}
\alias{stat.MW}
\title{stat.MW}
\usage{
stat.MW(Z, m)
}
\arguments{
\item{Z}{: pooled score vector with the first \eqn{m} components corresponding
to calibration observations and the last \eqn{n} components corresponding
to test observations.}

\item{m}{: calibration sample size.}
}
\value{
A rank vector.
}
\description{
It computes the ranks of the test observations in the
vector of calibration scores.
}
