% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g_old.R
\name{stats_G_j_MC}
\alias{stats_G_j_MC}
\title{stats_G_j_MC}
\usage{
stats_G_j_MC(N, g, B)
}
\arguments{
\item{N}{: pooled score sample size.}

\item{g}{: outlier density.}

\item{B}{: number of Monte Carlo repetitions used to estimate the test statistic.}
}
\value{
A vector of length N corresponding to the elementary statistics of Shiraishi test (1985).
}
\description{
stats_G_j_MC
}
