% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_higher.R
\name{var_analytical.Lk}
\alias{var_analytical.Lk}
\title{var_analytical.Lk}
\usage{
var_analytical.Lk(N, k, n)
}
\arguments{
\item{N}{: pooled sample size}

\item{k}{: an integer greater than 1, order of the Lehmann's alternative distribution}

\item{n}{: test sample size}
}
\value{
Asymptotic variance of Shiraishi (1985) test statistic
}
\description{
var_analytical.Lk
}
\examples{
v = var_analytical.Lk(N=100,k=2,n=50)
}
