% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g.R
\name{varG}
\alias{varG}
\title{varG}
\usage{
varG(n, m, stats_G_vector)
}
\arguments{
\item{n}{: test size}

\item{m}{: calibration size}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985). If NULL it will be computed in d_t using B_MC iterations}
}
\value{
A numeric value corresponding to the variance of the asymptotic distribution of Shiraishi test (1985)
}
\description{
varG
}
\examples{
g2 = function(x, k=2) ifelse(x<1 & x>0, k*x^(k-1), 0)
stats_G = stats_G_j_MC(N=1000, g=g2, B=10^3)
Var = varG(n=500, m=500, stats_G_vector = stats_G)

}
