% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_g_old.R
\name{varG}
\alias{varG}
\title{varG}
\usage{
varG(n, m, stats_G_vector)
}
\arguments{
\item{n}{: test size.}

\item{m}{: calibration size.}

\item{stats_G_vector}{: vector of elementary test statistics to perform the test in Shiraishi (1985).}
}
\value{
A numeric value corresponding to the variance of the asymptotic distribution of Shiraishi test (1985).
}
\description{
varG
}
